Провел рефактор проекта
