+++ .gitignore
@@ -1,4 +1,6 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+# folders
+.idea
+.vscode
 
 # dependencies
 /node_modules
@@ -17,7 +19,6 @@
 .env.development.local
 .env.test.local
 .env.production.local
-
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
