+++ package.json
@@ -1,15 +1,8 @@
 {
-  "name": "crypto_analize",
+  "name": "crypto_analyze",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.5.2",
-    "@types/node": "^16.18.54",
-    "@types/react": "^18.2.23",
-    "@types/react-dom": "^18.2.7",
     "bootstrap": "^5.3.2",
     "echarts": "^5.4.3",
     "react": "^18.2.0",
@@ -18,14 +11,11 @@
     "react-responsive-carousel": "^3.2.23",
     "react-router-dom": "^6.16.0",
     "react-scripts": "5.0.1",
-    "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "build": "react-scripts build"
   },
   "eslintConfig": {
     "extends": [
@@ -46,6 +36,10 @@
     ]
   },
   "devDependencies": {
-    "@types/echarts": "^4.9.19"
+    "@types/node": "^16.18.54",
+    "@types/react": "^18.2.23",
+    "@types/react-dom": "^18.2.7",
+    "@types/echarts": "^4.9.19",
+    "typescript": "^4.9.5"
   }
 }
