+++ src/actions/components/Charts/PieChart/PieChart.tsx
@@ -21,6 +21,7 @@ echarts.use([
     LabelLayout
 ]);
 
+// TODO: необходимо единнобразие в плане типизации проекта, потому что в одном месте type, а в другом interface
 export type PieChartOptions = {
     backgroundColor?: string;
     title: TitleComponentOption;
@@ -56,7 +57,7 @@ const PieChartComponent: React.FC<PieChartProps> = ({
             };
         }
     }, [chartRef, option]);
-
+    // TODO: по хорошему все стоит вынести в файл css, а не писать инлайновые стили
     return <div ref={chartRef} style={{ width, height }} />;
 };
