+++ src/actions/components/Charts/WaterfallChart/WaterfallChart.tsx
@@ -36,6 +36,8 @@ interface WaterfallChartProps {
     width?: string;
     height?: string;
 }
+// TODO: необходимо использовать единообразные ф-ции для компонентов, так как это может в будущем плохую шутку
+// TODO: либо expression, либо declaration
 
 const WaterfallChartComponent: React.FC<WaterfallChartProps> = ({
     option,
@@ -58,7 +60,7 @@ const WaterfallChartComponent: React.FC<WaterfallChartProps> = ({
             };
         }
     }, [chartRef, option]);
-
+    // TODO: по хорошему все стоит вынести в файл css, а не писать инлайновые стили
     return <div ref={chartRef} style={{ width, height }} />;
 };
