+++ src/actions/utils/post.ts
@@ -1,3 +1,5 @@
+// TODO: поменял расширение файла на .ts, так как в тут полностью отсуствует верстка
+
 export const fetchJsonData = async (url: string, method: string, requestData: any, headers: any) => {
     try {
         const response = await fetch(url, {
@@ -12,6 +14,7 @@ export const fetchJsonData = async (url: string, method: string, requestData: an
 
         const jsonData = await response.json();
         return jsonData;
+
     } catch (error) {
         throw error;
     }
@@ -31,6 +34,7 @@ export const fetchEncodedData = async (url: string, method: string, requestData:
         const headersData = await response.headers;
         console.log(response);
         return headersData;
+
     } catch (error) {
         throw error;
     }
